@@ -16,11 +16,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart' show timeDilation;
 
 class Photo extends StatelessWidget {
-  Photo({ Key key, this.photo, this.color, this.onTap }) : super(key: key);
+  Photo({ Key? key, required this.photo, this.onTap }) : super(key: key);
 
   final String photo;
-  final Color color;
-  final VoidCallback onTap;
+  final VoidCallback? onTap;
 
   Widget build(BuildContext context) {
     return Material(
@@ -43,15 +42,15 @@ class Photo extends StatelessWidget {
 
 class RadialExpansion extends StatelessWidget {
   RadialExpansion({
-    Key key,
-    this.maxRadius,
+    Key? key,
+    required this.maxRadius,
     this.child,
   }) : clipRectSize = 2.0 * (maxRadius / math.sqrt2),
        super(key: key);
 
   final double maxRadius;
   final clipRectSize;
-  final Widget child;
+  final Widget? child;
 
   @override
   Widget build(BuildContext context) {
@@ -74,7 +73,7 @@ class RadialExpansionDemo extends StatelessWidget {
   static const double kMaxRadius = 128.0;
   static const opacityCurve = const Interval(0.0, 0.75, curve: Curves.fastOutSlowIn);
 
-  static RectTween _createRectTween(Rect begin, Rect end) {
+  static RectTween _createRectTween(Rect? begin, Rect? end) {
     return MaterialRectCenterArcTween(begin: begin, end: end);
   }
 
@@ -134,7 +133,7 @@ class RadialExpansionDemo extends StatelessWidget {
                   pageBuilder: (BuildContext context, Animation<double> animation, Animation<double> secondaryAnimation) {
                     return AnimatedBuilder(
                       animation: animation,
-                      builder: (BuildContext context, Widget child) {
+                      builder: (BuildContext context, Widget? child) {
                         return Opacity(
                           opacity: opacityCurve.transform(animation.value),
                           child: _buildPage(context, imageName, description),
